Repaired SampleShrinkable
